brandbook-tokens/requirements.txt
anthropic>=0.28.0
PyMuPDF>=1.24.0
PyYAML>=6.0.1
